+++ MySQL/ty1_rtrh.js
@@ -93,41 +93,4 @@ function validateMessage(deviceInfo){
     return Joi.validate(deviceInfo, schema);
 }
 
-async function getNMinData(devID, minute){
-    const sqlQuery = `SELECT * FROM Device_${devType}_${devID} WHERE timestamp >= now() - INTERVAL ${minute} MINUTE;`
-    let connection;
-    try {
-      connection = await pool.getConnection();
-      result = await connection.query(`use ${database}`);
-      result = await connection.query(sqlQuery);
-      // console.log("Create DB", result);
-      return result;
-    } catch (ex) {
-      console.log("DB Error", ex.message);
-    } finally {
-      if (connection) connection.end();
-      console.log("Get N Minute Data Finally");
-    }
-}
-
-async function getLastNData(devID, qty){
-    const sqlQuery = `SELECT * FROM Device_${devType}_${devID} ORDER BY TIMESTAMP DESC LIMIT ${qty}`;
-    let connection;
-    try {
-      connection = await pool.getConnection();
-      result = await connection.query(`use ${database}`);
-      result = await connection.query(sqlQuery);
-      // console.log("Create DB", result);
-      return result;
-    } catch (ex) {
-      console.log("DB Error", ex.message);
-    } finally {
-      if (connection) connection.end();
-      console.log("Get Last N Data Finally");
-    }
-}
-
-
 exports.rtrhDbHandling = rtrhDbHandlings;
-exports.getNMinData_ty1 =getNMinData;
-exports.getLastNData_ty1 =getLastNData;
